pulled out moduleVisitor from no-unresolved, to be used with eslint-plugin-git
